refactor(shared): move internal/logging → shared/logging
Relocates rotator, styled handler, component logger, agent log, and
GUIDE.md. CLAUDE.md Routing section updated to the new GUIDE.md path.
