+++ internal/bot/retry_handler.go
@@ -5,7 +5,7 @@ import (
 	"log/slog"
 	"time"
 
-	"github.com/Ivantseng123/agentdock/internal/logging"
+	"github.com/Ivantseng123/agentdock/shared/logging"
 	"github.com/Ivantseng123/agentdock/internal/metrics"
 	"github.com/Ivantseng123/agentdock/shared/queue"
 )
