+++ internal/bot/workflow.go
@@ -13,7 +13,7 @@ import (
 	"github.com/Ivantseng123/agentdock/internal/config"
 	"github.com/Ivantseng123/agentdock/shared/crypto"
 	ghclient "github.com/Ivantseng123/agentdock/internal/github"
-	"github.com/Ivantseng123/agentdock/internal/logging"
+	"github.com/Ivantseng123/agentdock/shared/logging"
 	"github.com/Ivantseng123/agentdock/internal/mantis"
 	"github.com/Ivantseng123/agentdock/shared/queue"
 	slackclient "github.com/Ivantseng123/agentdock/internal/slack"
