+++ shared/logging/GUIDE.md
