+++ shared/logging/agent.go
