+++ shared/logging/agent_test.go
