+++ shared/logging/handler.go
