+++ shared/logging/handler_test.go
