+++ shared/logging/helpers.go
