+++ shared/logging/rotator.go
